Major refactoring - break MainService apart
This is part of a major effort to make the MainService less "essential"
to the operation of the app, since background services are really
frowned upon by Android 8+. We'll still need a background service for
some things, but at least things won't completely stop working if the
service is not running.

This has introduced lots of bugs and almost no testing has been
performed. Just dumping this here so it doesn't get lost.

"everything compiles"